siglent-sds: Rewrite eseries acquisition logic
This commit re-implements the whole acquisition logic for the Siglent
E-series (SDS1000X-E, SDS1000X-U, SDS2000X-E). The acquisition is
based on a state machine to keep the logic simple and robust.
Additionally this improves maintainability as the E-series acquisition
logic is now isolated from the older models which follow a partially
different logic and have other constraints when compared to the
E-series.